partial refactor: change factory to finder
